+++ src/contexts/NotificationContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { NotificationContextType } from "../types";
+
+export const NotificationContext = createContext<
+  NotificationContextType | undefined
+>(undefined);
