+++ src/contexts/ThemeContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { ThemeContextType } from "../types";
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined,
+);
