+++ src/contexts/UserContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { UserContextType } from "../types";
+
+export const UserContext = createContext<UserContextType | undefined>(
+  undefined,
+);
