+++ src/hooks/useUser.ts
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import { UserContext } from "../contexts";
+
+export const useUser = () => {
+  const context = useContext(UserContext);
+  if (!context) throw new Error("useUser must be used within a UserProvider");
+  return context;
+};
