+++ src/providers/ThemeProvider.tsx
@@ -1,0 +1,14 @@
+import { useState } from "react";
+import { ThemeContext } from "../contexts";
+
+export const ThemeProvider = ({ children }: React.PropsWithChildren) => {
+  const [theme, setTheme] = useState("light");
+  const toggleTheme = () =>
+    setTheme((prev) => (prev === "light" ? "dark" : "light"));
+
+  return (
+    <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
