+++ src/types/index.ts
@@ -1,0 +1,45 @@
+// 타입 정의
+export interface Item {
+  id: number;
+  name: string;
+  category: string;
+  price: number;
+}
+
+export interface User {
+  id: number;
+  name: string;
+  email: string;
+}
+
+export interface Notification {
+  id: number;
+  message: string;
+  type: "info" | "success" | "warning" | "error";
+}
+
+// ThemeContext 타입 정의
+export interface ThemeContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
+
+// UserContext 타입 정의
+export interface UserContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+// NotificationContext 타입 정의
+export interface NotificationContextType {
+  notifications: Notification[];
+  addNotification: (message: string, type: Notification["type"]) => void;
+  removeNotification: (id: number) => void;
+}
+
+// AppShell 타입 정의
+export interface Props {
+  items: Item[];
+  onAddItemsClick: () => void;
+}
